Alterando namespaces de interfaces e limpeza de codigo e importações.
