--- Restaurante.Domain/Comum/Modelos/IEntidade.cs
@@ -1,6 +1,0 @@
-﻿namespace Restaurante.Domain.Comum.Modelos
-{
-    public interface IEntidade
-    {
-    }
-}
