+++ Restaurante.Domain/Comum/Modelos/Intefaces/IEntidade.cs
@@ -1,0 +1,6 @@
+﻿namespace Restaurante.Domain.Comum.Modelos.Intefaces
+{
+    public interface IEntidade
+    {
+    }
+}
