+++ Restaurante.Infra/Usuarios/Encoder/Interfaces/IEncoder.cs
@@ -1,0 +1,9 @@
+﻿namespace Restaurante.Clientes.Infra.Usuarios.Encoder.Interfaces
+{
+    public interface IEncoder
+    {
+        string Encode(string encode);
+        string Decode(string decode);
+    }
+
+}
