+++ Restaurante.Infra/Usuarios/Encoder/Interfaces/IPasswordEncoder.cs
@@ -1,0 +1,7 @@
+﻿namespace Restaurante.Clientes.Infra.Usuarios.Encoder.Interfaces
+{
+    public interface IPasswordEncoder : IEncoder
+    {
+        bool VerficarSenha(string senha, string senhaEnconder);
+    }
+}
