+++ Restaurante.Infra/Usuarios/Encoder/SenhaEncoder.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Restaurante.Clientes.Infra.Usuarios.Encoder.Interfaces;
+using System;
 
 namespace Restaurante.Clientes.Infra.Usuarios.Encoder
 {
